style(mm, task, trap): apply cargo clippy suggestions

+++ os/src/trap/mod.rs
@@ -131,7 +131,7 @@ fn set_kernel_trap_entry() {
 
 fn set_user_trap_entry() {
     unsafe {
-        stvec::write(TRAMPOLINE as usize, TrapMode::Direct);
+        stvec::write(TRAMPOLINE, TrapMode::Direct);
     }
 }
